# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Email Assistant API
  version: 1.0.0
  description: |
    REST API for the Email Assistant used by the Influencer Management System.
    Notes:
    - Write endpoints require an API key in header `X-API-Key` when configured.
    - OAuth web flow is provided to bind Gmail sender accounts.
servers:
  - url: /
    description: Relative server (behind your domain/proxy)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
...
    SendEmailsRequest:
      type: object
      required: [sender_email, excel_file_path]
      properties:
        sender_email:
          type: string
          format: email
        excel_file_path:
          type: string
        subject:
          type: string
        content:
          type: string
        html_content:
          type: string
        attachments:
          type: array
          items:
            type: string
          description: Attachment file IDs under files/
        start_time:
          type: string
          format: date-time
        min_interval:
          type: integer
          default: 20
        max_interval:
          type: integer
          default: 90
    SendEmailsResponse:
      type: object
      properties:
        success:
          type: boolean
        total:
          type: integer
        success_count:
          type: integer
        failed_count:
          type: integer
        duration:
          type: string
        start_time:
          type: string
        end_time:
          type: string
    TemplateSendRequest:
      type: object
      required: [sender_email, excel_file_path]
      properties:
        sender_email:
          type: string
          format: email
        excel_file_path:
          type: string
        attachments:
          type: array
          items:
            type: string
        start_time:
          type: string
          format: date-time
        min_interval:
          type: integer
          default: 20
        max_interval:
          type: integer
          default: 90
    TemplateSendResponse:
      type: object
      properties:
        success:
          type: boolean
        stats:
          type: object
        duration:
          type: string
        status:
          type: string
...
paths:
...
  /api/authenticate:
    post:
      tags: [OAuth]
      summary: Authenticate sender (legacy/local)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sender_email]
              properties:
                sender_email:
                  type: string
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /oauth/google/authorize:
    get:
      tags: [OAuth]
      summary: Start Google OAuth (redirect)
      parameters:
        - in: query
          name: sender_email
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to Google consent page
  /oauth/google/callback:
    get:
      tags: [OAuth]
      summary: OAuth callback (exchange code for token)
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Binding result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  email:
                    type: string
  /api/validate_excel:
    post:
      tags: [Utility]
      summary: Validate Excel file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [excel_file_path]
              properties:
                excel_file_path:
                  type: string
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/preview:
    post:
      tags: [Utility]
      summary: Preview pending emails from Excel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [excel_file_path]
              properties:
                excel_file_path:
                  type: string
      responses:
        '200':
          description: Preview data
          content:
            application/json:
              schema:
                type: object
  /api/send_emails:
    post:
      tags: [Send]
      summary: Send emails (classic mode)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendEmailsRequest'
      responses:
        '200':
          description: Send result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendEmailsResponse'
        '401':
          description: Unauthorized (missing/invalid API key)
  /api/send_template_emails:
    post:
      tags: [Send]
      summary: Send emails using templates
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateSendRequest'
      responses:
        '200':
          description: Send result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateSendResponse'
        '401':
          description: Unauthorized (missing/invalid API key)
...
  /api/pause:
    post:
      tags: [Control]
      summary: Pause sending
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/resume:
    post:
      tags: [Control]
      summary: Resume sending
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/stop:
    post:
      tags: [Control]
      summary: Stop sending
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/statistics:
    post:
      tags: [Utility]
      summary: Get statistics from Excel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [excel_file_path]
              properties:
                excel_file_path:
                  type: string
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                type: object
  /api/validate_templates:
    post:
      tags: [Templates]
      summary: Validate templates against Excel columns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [excel_file_path]
              properties:
                excel_file_path:
                  type: string
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
  /api/preview_template_emails:
    post:
      tags: [Templates]
      summary: Preview template-rendered emails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [excel_file_path]
              properties:
                excel_file_path:
                  type: string
                max_previews:
                  type: integer
                  default: 3
      responses:
        '200':
          description: Preview result
          content:
            application/json:
              schema:
                type: object
...